turris-maintain: fix reading of IP addresses
Issue foris-controller#146 introduced new way of storing IP address and
netmask for lan, to be more aligned with the way LuCI stores it.

However it breaks several maintenance scripts, which in turn won't send
notification to re/foris to show the "reloading network..." spinner.

This fixes reading of IP address from uci.