reforis: add HOST_BUILD_DEPENDS for node
It happened that there was some issues related to node/host, which was
not compiled and even reForis itself was somehow compiled, it is
required for reForis JS and its translations.

This ensures that node/host will be for both PKG_BUILD_DEPENDS and
HOST_BUILD_DEPENDS as it is exactly done in node feed, which we are
using. This could explain why node packages failed, but reForis didn't.