Revert "Revert "lighttpd: start Foris and reForis FastCGI backends on demand""
This reverts commit 14abc044af6076364b015610d082e0aee085e9e8.

The issue with reforis starting was that both reforis and lighttpd were
creating FCGI socket. Thus lighttpd created socket and send request to
it while reForis replaced that socket when it initialized and thus never
get the initial request. The removal of socket configuration solves the
issue and the patch added here is that fix.