turris-maintain: split off turris-backup and rewrite
This makes backup scripts more accessible and also adds additional
features. The short list is:
* Inclusion of sysupgrade files (OpenWrt's way to mark files for backup)
* Optional inclusion of modified configuration files
* More strict inclusion for previously included files
* Protection against recovery from other major versions of OS

The last change is the most important as it is not just plain addition
but the incompatible change. This is intentional as the protection is
more desirable than the compatibility. The use of the scripts is same as
the previously but recovery of backup from different major OS version is
not allowed.