updater-ng: backport two patches
The second patch is fix. This backports patch that fixes issues with
scripts terminating but not correctly closing all outputs results into
hang.

The first patch is a feature. The turris-build repository contains
second part of it and it is already merged. It is not desirable to
release new version of updater because of other changes thus backporting
only this patch is a good solution.