tos3to4-notify: enable automatic migration
This enables automatic migration for all Turris 1.x routers running on
BTRFS and up to two thousands of Omnias (the exact number depends if we
hit already migrated routers or not). That should give us up to three
thousands of routers to handle and support migration on.

The Omnia routers are selected based on serial number. The decimal
representation of serial number has to end with zero. The decimal
representation can be obtained using (package bc has to be installed):
  echo "ibase=16; $(atsha204cmd serial-number)"| bc