tos3to4: new package tos3to4-earliest
In the past the tos3to4-early was installed alongside of updater but
that wasn't ideal so it was moved to be installed later with latest
version of updater after the first replan. This created issue that
script that should allow newest updater to run wasn't installed and
executed. At the same time we do not want to install tos3to4-early due
to its dependencies that we can't simply overcome in the old updater.
The alternative is to simply introduce tos3to4-earliest that contains
the simple localrepo removal script. This is going to be dependency of
updater and thus installed before the new version is installed.