Revert "turris-timezone: temporary fix for l18n"
l18n developers did some Python magic, because with a recent update to
2021.3.1 version, they add Python 3.10 compatibility, which
unfortunately requires having an installed wheel before running setup.py.
Also, turris-timezone requires l18n otherwise it is not compiled.

The issue was with OpenWrt 19.07, which uses always the latest version for
host(side) packages, and luckily for us Python OpenWrt maintainers
(commodo, jefferyto) added the option to have pinned host dependencies for OpenWrt since 21.02.

This is done in the packages feed and lang/host-pip-requirements folder.

This reverts commit a6670e6c5366b4dd6f64a95f1ef195c773e28bd4.