tos3to4: fix invalid VLAN assignment
The `vid` is in default set by LuCI but if user configures stuff
directly it can use only `vlan` and in default that is used as `vid`.
This is probably a common way most of the advanced users configured
switch. We need to read vlan and use that as default to vid.