fosquitto: raise respawn_timeout
Raise respawn_timeout from 0.2 to 2s in order to not to try respawn
process too fast.

Shorter timeout caused issue on certain variants of Mox, that fosquitto
was started too soon during boot and due to the short timeout, it
reached maximum number of respawns.

Which in turn broke foris-controller, which depends on fosquitto, and
both services needed to be restarted manually.