fix: network-devices cover ifname as option
The config_list_foreach does not read the real value and instead it only
works with values set in uci as list. The issue is that OpenWrt uses
option for ifnames as well so we have to support that.
This is not nice but we simply have to expand the list of those devices
to add them to the list.