sentinel-fwlogs: update to version 0.3.0
Since version 0.3.0, it is required dependency logc-czmq. [1]

We need to switch to dist tarball, otherwise there is no configure.
Also, we need to avoid using ``PKG_FIXUP:=autoreconf`` for now, since it
requires to use fwrecorder, which is development tool and it can not be
disabled by using ``CONFIGURE_ARGS``.

[1] https://gitlab.nic.cz/turris/sentinel/fwlogs/-/blob/e132a6e4e1c86e4b42d637c7f22ac9b66ec64b2b/CHANGELOG.md#030-2022-03-10

[switch to dist tarball, use autotools.mk]
Co-authored-by: Karel Kočí <karel.koci@nic.cz>