l10n-supported: rename nb to nb_NO
We are using Weblate for translations and this system uses and suggests
the nb_NO and not just plain nb. The effect is that we have to do move
of the translation every time the new translation of nb_NO is created
(Weblate then understands it and uses nb). This si not clean at all.
Thus this move should make it way more systematic.