l10n-supported: remove and add few languages
It is some time since we updated the list of languages. We want to show
only languages that have reasonable activity (that they have reasonable
number of translated strings). Our limit here is 10% and inclusion in
most of our projects.

The following languages were removed because they lack the activity and
the percentage of translated strings dropped to single digits:
* Danish
* Italian
* Lithuanian

The following languages are added because they have reasonable amount of
translations:
* Spanish
* Swedish
* Dutch