treewide: use number prefixes for lighttpd configuration files
The lighttpd now prefixes the configuration files with numbers to ensure
their ordering. We should use this as well and this changes names of all
our lighttpd configuration files to be prefix with number.

The rules used here is:
80: any regular web service
70: dependency of other web services (just as API not for configuration)
60: root configuration
50: dependencies for other services (variables required for configuration)
45: the conditional modification of basic Lighttpd behavior
40: the SSL enable (modification of basic behavior)

turris-root.conf was prefixed by 90 instead of 60 because it appends "/"
mapping to alias.url which prevents succeeding mappings from matching.

Co-Authored-By: Šimon Bořek <simon.borek@nic.cz>