turris1x-uboot: Introduce new U-boot versions
This package contains two new development U-boot versions for Turris
1.x. Only one is shipped with this package to users as the other
  version is nice to have if something goes wrong.

These files were provided by @prohar, who is working to have Turris 1.x
upstream support in U-boot and kernel as well. While working on it, he
found some serious issues in U-boot for mpc85xx, which are currently on
review in their mailing list.

This could be helpful for experienced users, who might try to test it
and let us know if all the things work as they should.

This was compile tested for Turris 1.x.