mox-a53-firmware: fix value in rshift
The total size of a53-firmware partition is 1441792 bytes (0x160000).
The first 0x160000 bytes is BL1 and what remains is FIT.

On the absolute position 131184 (131072 + 112), there is offset, where
begins U-boot and on the position 131192 (131072 + 120), there you can
find the total size of U-boot.
(= Total size of partition - size BL1 - offset U-boot)

Reported-by: Pali Rohár <pali@kernel.org>
Co-authored-by: Pali Rohár <pali@kernel.org>