turris1x-uboot: compile it from source code
Pali Rohár <pali@kernel.org> sent many improvements for PowerPC in
U-boot and Kernel Linux as well. Also, he added support for Turris 1.x
routers, but it is currently on review. We would like to have some
development binaries available to be able to test it in Turris OS.

In CZ.NIC Turris U-boot repository, since today there is a published branch
turris1x-v2022.04, which currently has this commit as head
16fa34e9bc60af50759c389ce40ae35322fdbf55 ("powerpc: Add support for
CZ.NIC Turris 1.x routers"), which is used to compile U-boot binaries
for NOR and SDCard.