lighttpd: run it as root
Upstream solution in OpenWrt uses http as user and www-data as group,
unfortunately, this breaks new Turris OS installations.

When I used the upstream solution inside init script to create http
user and the group as well, then it does not work for us and we need to
release Turris OS 6.0 fast.

Fixes:
1. 2022-10-14 16:35:22: (../src/server.c.1410) can't find username http
2. 2022-10-14 16:32:48: (../src/configfile.c.1720) opening errorlog '/var/log/lighttpd/error.log' failed: Permission denied
2022-10-14 16:32:48: (../src/server.c.1584) Opening errorlog failed. Going down.

Fixes: 23f5699469a6c0970de795be4ae4e05e0451527e ("lighttpd &
turris-webapps: prepare to support upstream OpenWrt lighttpd pkg")